# Complete FileSystemService integration and architectural cleanup